fix solution, and cleanup the code
