+++ code/lib/core-server/src/utils/constants.ts
@@ -1,0 +1,1 @@
+export const DEBOUNCE = 100;
