+++ code/lib/core-server/src/utils/getAccessControlMiddleware.ts
@@ -1,0 +1,15 @@
+export function getAccessControlMiddleware(crossOriginIsolated: boolean) {
+  return (req, res, next) => {
+    res.header('Access-Control-Allow-Origin', '*');
+    res.header('Access-Control-Allow-Headers', 'Origin, X-Requested-With, Content-Type, Accept');
+    // These headers are required to enable SharedArrayBuffer
+    // https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/SharedArrayBuffer
+    if (crossOriginIsolated) {
+      // These headers are required to enable SharedArrayBuffer
+      // https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/SharedArrayBuffer
+      res.header('Cross-Origin-Opener-Policy', 'same-origin');
+      res.header('Cross-Origin-Embedder-Policy', 'require-corp');
+    }
+    next();
+  };
+}
