+++ code/lib/core-server/src/utils/router.ts
@@ -1,0 +1,5 @@
+import { Router } from 'express';
+
+// @ts-expect-error (Converted from ts-ignore)
+
+export const router: Router = new Router();
