+++ code/lib/core-server/src/utils/versionStatus.ts
@@ -1,0 +1,7 @@
+import type { VersionCheck } from '@storybook/types';
+
+export const versionStatus = (versionCheck: VersionCheck) => {
+  if (versionCheck.error) return 'error';
+  if (versionCheck.cached) return 'cached';
+  return 'success';
+};
